Expose FilterSparseFields to allow API developers to implement jsonapi.org Sparse Fieldsets handling.
